Move logic into api.py in preparation for refactor
